subpictures: add screen size field
Add a scren size field which describes the size for which the SPU has
been prepared just before being used by the output. It is useful in case
of external scaling for the subpictures, in case the vout display is
able to rescale them itself in a more efficient way.