qt: add qt-fullscreen-to to replace qt-fullscreen-screennumber
Screen number is a leftover of Xinerama and was hacked into the list of
Qt screens which provided constant identifiant, but there are not
user-friendly to use.
Screen name are providing a much more reliable and user-friendly way to
set the screen and are what vout window API expects too.