opengl: prevent opengl calls from different threads
Most display controls, especially sizing ones, have been made synchronous,
while an OpenGL context should be used in only one thread at a time and
should not interfere with the context of the control calling thread.
It is especially harmful when the calling context is using OpenGL, like
in interfaces.