window: add has_shared_egl info flag
UI integration of the video might need shared display, for instance with
wayland. When the display connection is shared, calling eglTerminate
will destroy every other context. The `has_shared_egl` flag prevents
this by notifying the possible EGL clients of the window that they
should not terminate the EGLDisplay themselves.