egl: tune eglTerminate behaviour with display_reference availability
EGL_KHR_display_reference specifies the behaviour of eglTerminate
with regards to reference counting. In addition, it specifies that
the default is that reference counting is disabled X11 or wayland,
but enabled on Android.

As we can prevent EGL GL provider from terminating its state, it
makes sure the state is correctly refcounted in any case.