contrib: soxr: don't enforce CMAKE_SYSTEM_NAME on android
The CMake script in soxr will try to infer the current system and
setting the CMAKE_SYSTEM_NAME variable prevents CMake from finding which
system it really is and believe it's Windows. Because PKG_CONFIG_PATH
entries are separated by a semi-colon on Windows, it prevents finding
libraries like libavcodec.

This commit disable correct SIMD detection if CMAKE_ANDROID_ARCH_ABI is
not correctly set or if CMAKE_SYSTEM_PROCESSOR has an invalid value with
regard to android autodetection scripts in CMake
(Android-Determine.cmake).