libvlc: media: add parse_associated_media option
This option will allow parsing other media items associated to the
current media through the use of `libvlc_media_player_add_slave` or
similar other mechanism. It is not enable by default as being more IO
intensive and giving a different behaviour than the previous one.

It is needed in particular to find the different es_id associated with
the different tracks in a LibVLC client wanting to make use of the
unstable stream output options, and wanting to use filters like
duplicate with `select-*` option.