transcode: video: append converter before encoder if needed
The encoder is able to override its input format, especially to change
the input chroma. The semantic is that the encoder user should either
change the input format or refuse the module probing.

The transcode video pipeline had a dedicated filter_chain to handle such
conversion but it was never initialized.

Store encoder input format before it is initialized, and initialize the
final filter chain if the format is different from what the encoder
requested.

Fix #22703