libvlc: let the OpenGL host app tell libvlc its rendering area changed
VLC adjusts the text rendering based on the rendering area size, so we should
be able to adapt to changes in the host app.

This is using the wextern dummy window, like it's done with D3D callbacks. The
callback is given to the host app when the window is enabled and reset when the
window is disabled.

Not sure how it's supposed to be done with Android but given it uses a
regular window, it probably reports size changes already.