configure: split dynamic plugins from dynamic loader
As we can embed static and dynamic plugins at the same time, it makes
sense to build LibVLC-provided modules inside VLC but still allow
external out-of-tree plugins to be added later.

It allows extending LibVLC through the dynamic loader while keeping the
official binaries, which make sense on iOS and Android. Indeed, on those
platform, shared library are either limited (support since iOS 8/9 on
one side and limit to 128 shared objects including system libraries on
the other side until API 23 [1] thanks to patch [2]) or with a
non-negligeable performance cost.

It is also much better than providing a plugin interface in which you
merge all modules from LibVLC and force the use of the dynamic loader as
the current inheritance mechanism doesn't play nicely with such
construction, and current buildsystem for Android and iOS are conceived
for static builds.

[1] https://android.googlesource.com/platform/bionic/+/android-n-mr2-preview-1/android-changes-for-ndk-developers.md
[2] https://android.googlesource.com/platform/bionic/+/ba98d9237b0eabc1d8caf2600fd787b988645249%5E%21/