threads: store the condition variable value as an atomic_uint with LIBVLC_NEED_CONDVAR
We can include stdatomic.h in the headers, we do it in many places. But we
can't use it with C++ because the std:atomic_int storage is likely different
and can be initialized/released the same way.

We change the static assert in the core to make sure the type in the union that
will be used for storage by C++ modules is correct.

In C++11 there is std::condition_variable anyway. The vlc_cond_t would only be
useful to share it with C code in a module or legacy C++ code.