placebo_utils: unexport private API
This API is only used in modules and compiled in every modules using
placebo (opengl common sources). In the end, with the OpenGL refactor,
it will probably be confined into its own plugin so there is no need to
export it. In addition, the symbols are not even exposed in
libvlccore.sym.