apple: build: export host tools in hostenv wrapper
Avoid exporting environment variables as it prevents using a different
native compiler than apple one for extra tools. The tools can be fetched
directly during script startup and are in VLC_ prefixed variables to
avoid conflicts and follow the script conventions.

Note that native compiler can be choosen using CC= for extra/tools as
a side effect but this is not the goal of this patch.