contrib: use makefile instead of environment variable
MacOSX builds and iOS builds need fixups for the autoconf function
lookup as functions might be available with newer SDK but not available
with the target SDK currently supported, and autoconf doesn't detect
that correctly. These fixups are made through environment variables
declared in extras/package/macosx/env.build.sh. They are also needed
both for contribs and VLC itself.

As long as reconfiguration doesn't happen, it is possible to compile
VLC using make without sourcing the script. However the contribs are
not storing those exports and might fail during the middle of build
if the sourcing has been forgotten.

This patch allows building the contrib with make directly, without
needing to call the build script again or redefine the correct
environment variables before.