android: refactor SurfaceTexture handling
Refactor SurfaceTexture related API into a structure with vtable, so as
to implement both NDK and JNI API as separate vtable.

It will be exposed to modules and will allow creating SurfaceTexture
directly in the module needing them, instead of relying on an external
SurfaceTexture provided by the binding.

jsurface and ANativeWindow are also both exposed to the user since we
don't expose the ANativeWindow/JNI functions.

The object is made so that SurfaceTexture are always created in the
detached state.