WIP: mediacodec: create SurfaceTexture in decoder
Initially, the SurfaceTexture reference was created through
AWindowHandler and typically was obtained through a TextureView. By
creating the SurfaceTexture ourselves, we can avoid requesting Android
libVLC user to use libvlc_media_player_set_android_context even if it
will only decode video and we don't rely on code that is not in vlc
core.

In contrary to TextureView, the SurfaceTexture obtained that way will
only use a single buffer.